chatbot-backend/requirements.txt
fastapi
uvicorn
requests
python-dotenv
google-generativeai
google-auth

